Fixed two bugs
LiveSplit no longer instantly splits when the timer starts and a level was completed while the timer wasn't running.

The death counter no longer gets added to the layout regardless of whether the setting is enabled or not. If the setting is unchecked (and was checked before, so the text field is present), the text field still needs to be deleted manually.